# Improve efficiency of merging bulk insertions into the hash index